pip-requirements.txt
Pillow==2.4.0
PyOpenGL==3.0.2
numpy==1.8.1
wsgiref==0.1.2
